Update patch set 13
Patch Set 13:

We would like to kindly ask for your assistance in helping us merge the above patch, which is currently blocking DLUX project from delivering Carbon activities.  More detailed description of the blocking issues can be found on the weather and mailing list.  Please take a moment to +2 and press the blue "Submit" button, or if you have additional concerns or inquiries, please do not hesitate to let us know.  We will be tracking the blocking issues on the Google Tracking Spreadsheet.

Please note that projects must build successfully in autorelease by M4 timeframe for the Carbon Simultaneous Release.  The TSC reserves the right to remove projects from Carbon if not building successfully in autorelease by RC0.  Please do not hesitate to contact us if you have any questions or concerns.

Patch-set: 13